Add embeds

+++ src/main/java/com/github/echo/core/commands/admin/UpdateMessageCommand.java
@@ -4,47 +4,44 @@ import com.github.echo.core.Main;
 import com.github.echo.core.commands.Command;
 import com.github.echo.core.commands.Modules;
 import com.github.echo.utilities.Constants;
+import com.github.echo.utilities.DefaultEmbeds;
 import com.github.echo.utilities.MessageUtilities;
+import com.github.echo.utilities.RoleReactEmbeds;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.awt.*;
-import java.io.File;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class LiveCommand extends Command {
+public class UpdateMessageCommand extends Command {
 
-    private static final String link = "https://twitch.tv/scorched-echo";
-    private static final String banner = "https://media.discordapp.net/attachments/888533896952676353/888534241137287259/Banner_with_Bat.png?width=2052&height=843";
 
     @Override
     protected void onCommand(MessageReceivedEvent mre, String[] args) {
-        Main.getLog().info("LIVE (called by " + mre.getAuthor().getAsTag() + ")");
+        Main.getLog().info("UPDATE MESSAGE (called by " + mre.getAuthor().getAsTag() + ")");
 
-        // handle args
-        StringBuilder combinedArgs = new StringBuilder();
-        args[0] = "";
-
-        for (String x : args) {
-            combinedArgs.append(x).append(" ");
+        String userID = mre.getAuthor().getId();
+        if (!userID.equals(Constants.OWNER_ID)) {
+            mre.getChannel().sendMessage("This command can only be used by the owner.").queue();
+            return;
         }
-        String title = combinedArgs.toString().replaceFirst(" ", "");
 
-        Main.getLog().debug("Going Live with: " + title, Constants.stageCommand);
+        if (args.length < 2) {
+            mre.getChannel().sendMessage("Please define a message ID.").queue();
+        }
 
-        EmbedBuilder embed = new EmbedBuilder();
-        MessageUtilities.addEmbedDefaults(embed);
-        embed.setColor(new Color(111, 83, 112));
+        /*List<TextChannel> channels = mre.getGuild().getTextChannels();
 
-        embed.setTitle("scorchedE.C.H.O is going live!", link);
-        embed.setImage(banner);
+        for (TextChannel channel : channels) {
+            channel.
+        }*/
 
-        embed.addField(title, "Click the title to join!", false);
-
-        mre.getChannel().sendMessage("@everyone").queue();
-        mre.getChannel().sendMessageEmbeds(embed.build()).queue();
+        mre.getGuild().getTextChannelById("935321062583980044").retrieveMessageById("935344475511865344").queue((message -> {
+            message.editMessageEmbeds(DefaultEmbeds.marketRules()).queue();
+        }));
 
         // delete msg
         mre.getMessage().delete().queue();
@@ -53,7 +50,7 @@ public class LiveCommand extends Command {
 
     @Override
     public List<String> getAliases() {
-        return Arrays.asList("live", "twitch");
+        return Arrays.asList("umsg", "updatemsg");
     }
 
     @Override
